requirements.txt
Django==1.11.7
django-appconf==1.0.3
django-compressor==2.3
django-environ==0.4.5
djangorestframework==3.6.3
psycopg2-binary==2.7.6.1
python-memcached==1.59
pytz==2019.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.12.0
whitenoise==4.1.2
xlwt==1.3.0
